move `DefaultValue` component to `@graphiql/react`
