+++ packages/graphiql-react/src/explorer/components/default-value.css
@@ -1,0 +1,3 @@
+.graphiql-doc-explorer-default-value {
+  color: var(--color-green);
+}
